# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sosyal Medya API",
        "description": "backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/follow/{user}": {
            "post": {
                "tags": [
                    "Follow"
                ],
                "summary": "Takip et / Takipten çık",
                "operationId": "887413cacc3bdc85e7baf14c53c74f58",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{user}/followers": {
            "get": {
                "tags": [
                    "Follow"
                ],
                "summary": "Takipçiler",
                "operationId": "1e4444d2ba3f2ed144b414d735608dd8",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Takipçi listesi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{user}/following": {
            "get": {
                "tags": [
                    "Follow"
                ],
                "summary": "Takip edilenler",
                "operationId": "2640a5fed728395709cb4b3a1e1b9785",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Takip edilenler listesi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{post}/like": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Beğen / beğeniyi kaldır",
                "operationId": "0a6f1b02103f3da9c78f44153cb6fb11",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "liked": {
                                            "type": "boolean"
                                        },
                                        "likes_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/posts/{post}/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Beğenenler listesi",
                "operationId": "fc3aee049d33ddc402e4d7d1f689c58d",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Beğenenler"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Bildirimlerim",
                "operationId": "6d56a555955089a73fdf3f7229e8ead8",
                "responses": {
                    "200": {
                        "description": "Bildirimler listesi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Okunmamış bildirim sayısı",
                "operationId": "f3e05b69d516049be949f02a45e0eedd",
                "responses": {
                    "200": {
                        "description": "Okunmamış sayısı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "unread_count": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": {
        "1": {
            "sanctum": []
        }
    }
}